complited Product, ProductCRUD

+++ src/main/java/com/artShop/SQL/Product.java
@@ -4,7 +4,7 @@ public class Product {
     private String productCode;
     private String name;
     private String description;
-    private double price;
+    private int price;
     private int amount;
 
     public String getProductCode() {
@@ -31,11 +31,11 @@ public class Product {
         this.description = description;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 
@@ -48,7 +48,7 @@ public class Product {
     }
 
     public Product(String productCode, String name,
-                   String description, double price, int amount) {
+                   String description, int price, int amount) {
         this.productCode = productCode;
         this.name = name;
         this.description = description;
